feat(ui): add new character card and update party roster
