+++ client/src/components/CharacterCard.module.css
@@ -1,0 +1,56 @@
+.characterCard {
+  background: #161d29;
+  border-radius: 20px;
+  box-shadow: 0 2px 24px #000a;
+  padding: 1.6rem 1.5rem 1.1rem;
+  color: #ede8d8;
+  width: 320px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: relative;
+}
+.portrait {
+  width: 90px;
+  height: 90px;
+  border-radius: 50%;
+  border: 4px solid #393e53;
+  object-fit: cover;
+  margin-bottom: 12px;
+  margin-top: -36px;
+}
+.badge {
+  position: absolute;
+  right: 20px;
+  top: 20px;
+  padding: 6px 14px;
+  border-radius: 12px;
+  font-size: 1rem;
+  font-weight: 600;
+  color: #fff;
+  background: var(--badge-color, #c53d38);
+}
+.name {
+  font-family: 'Merriweather', serif;
+  font-size: 2rem;
+  font-weight: bold;
+  margin-top: 8px;
+}
+.subtitle {
+  color: #b6bace;
+  font-size: 1.1rem;
+  margin-bottom: 8px;
+}
+.desc {
+  font-size: 1.12rem;
+  text-align: center;
+  margin: 14px 0;
+}
+.hpBar {
+  margin-top: 16px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  font-size: 1.18rem;
+  color: #e6504d;
+}
